feat: разделит комплексные файлы команд и запросов на составляющие
Разделит на Command, Handler и Validator.
