+++ BankAccounts.Api/Features/Transactions/Commands/PerformTransfer/PerformTransferHandler.cs
@@ -5,48 +5,19 @@ using BankAccounts.Api.Features.Transactions.Dtos;
 using BankAccounts.Api.Infrastructure.CurrencyService;
 using BankAccounts.Api.Infrastructure.Repository.Accounts;
 using BankAccounts.Api.Infrastructure.Repository.Transactions;
-using FluentValidation;
-using MediatR;
+
 // ReSharper disable once UnusedType.Global Класс используется посредником
 
-namespace BankAccounts.Api.Features.Transactions.Commands;
+namespace BankAccounts.Api.Features.Transactions.Commands.PerformTransfer;
 
 /// <summary>
-/// Произвести перевод средств между счетами
-/// </summary>
-public static class PerformTransfer
-{
-    /// <summary>
-    /// Команда для проведения трансфера
-    /// </summary>
-    public record Command : IRequest<TransactionDto>
-    {
-        /// <summary>
-        /// Id пользователя
-        /// </summary>
-        public Guid OwnerId { get; set; }
-        /// <summary>
-        /// Id исходящего счета
-        /// </summary>
-        public int FromAccountId { get; init; }
-        /// <summary>
-        /// Id счета назначения
-        /// </summary>
-        public int ToAccountId { get; init; }
-        /// <summary>
-        /// Сумма денежных средств
-        /// </summary>
-        public decimal Amount { get; init; }
-    }
-
-    /// <summary>
     /// Обработчик команды. Трансфер происходит с использованием двух транзакций, одна снимает средства с исходного счета
     /// и одна зачисляет на конченый счет.
     /// </summary>
-    public class Handler(IAccountsRepositoryAsync accountsRepository, ITransactionsRepositoryAsync transactionsRepository, ICurrencyService currencyService, IMapper mapper) : BaseRequestHandler<Command, TransactionDto>
+    public class PerformTransferHandler(IAccountsRepositoryAsync accountsRepository, ITransactionsRepositoryAsync transactionsRepository, ICurrencyService currencyService, IMapper mapper) : BaseRequestHandler<PerformTransferCommand, TransactionDto>
     {
         /// <inheritdoc />
-        public override async Task<TransactionDto> Handle(Command request,  CancellationToken cancellationToken)
+        public override async Task<TransactionDto> Handle(PerformTransferCommand request,  CancellationToken cancellationToken)
         {
 
             var fromAccount = await GetValidAccount(accountsRepository, request.FromAccountId, request.OwnerId, cancellationToken);
@@ -84,19 +55,3 @@ public static class PerformTransfer
             return mapper.Map<TransactionDto>(transactionFrom);
         }
     }
-    /// <summary>
-    /// Валидатор команды
-    /// </summary>
-    public class CommandValidator : AbstractValidator<Command>
-    {
-        
-        /// <inheritdoc />
-        public CommandValidator()
-        {
-            RuleFor(command => command.OwnerId).NotEmpty();
-            RuleFor(command => command.FromAccountId).GreaterThan(0).NotEqual(command => command.ToAccountId);
-            RuleFor(command => command.ToAccountId).GreaterThan(0);
-            RuleFor(command => command.Amount).GreaterThan(0);
-        }
-    }
-}
