+++ BankAccounts.Api/Features/Transactions/Queries/GetBankStatement/AccountOperation.cs
@@ -1,0 +1,18 @@
+﻿namespace BankAccounts.Api.Features.Transactions.Queries.GetBankStatement;
+
+/// <summary>
+/// Операция по счету
+/// </summary>
+/// <param name="DateTime">Время проведения</param>
+/// <param name="CounterPartyId">Счет получателя (если есть)</param>
+/// <param name="Sum">Сумма</param>
+/// <param name="AfterBalance">Баланс после операции</param>
+/// <param name="Description">Описание</param>
+// ReSharper disable NotAccessedPositionalProperty.Global Свойства используются
+public record AccountOperation(
+    DateTime DateTime,
+    int CounterPartyId,
+    decimal Sum,
+    decimal AfterBalance,
+    string? Description
+);
